Extract toolbar view
