+++ RvChecklist/Views/Checklist/ChecklistView.swift
@@ -27,7 +27,7 @@ struct ChecklistView: View {
             .edgesIgnoringSafeArea([.top])
             .toolbar {
                 ToolbarItem(placement: .primaryAction) {
-                    ChecklistToolbar(category: $category)
+                    ToolbarView(category: $category)
                 }
             }
         }
@@ -49,57 +49,6 @@ struct ContentView_Previews: PreviewProvider {
                     .environmentObject(ModelData())
                     .previewDevice(PreviewDevice(rawValue: deviceName))
                     .previewDisplayName(deviceName)
-            }
-        }
-    }
-}
-
-
-struct ChecklistToolbar: View {
-    
-    @EnvironmentObject var modelData: ModelData
-    @Binding var category: String
-
-    var body: some View {
-        Menu {
-
-            Section(header: Text("List Type")) {
-                //TODO: map/reduce unique list of categories
-                ForEach(Array(Set(modelData.checklist.map { $0.category })), id: \.self) { list in
-                    Button(list) {
-                        //
-                        category = list
-                    }
-                }
-            }
-
-            Section {
-                Button("Start New Trip") {
-                    startNewTrip()
-                }
-            }
-
-        } label: {
-            Label("", systemImage:"ellipsis.circle.fill")
-        }
-        .foregroundColor(.white)
-        .font(.subheadline)
-    }
-    
-    func startNewTrip() {
-        //TODO: display screen to enter name, date, etc.
-        
-        // then clear isDone on all trip items
-        clearChecklist(category: "Pre-Trip")
-        clearChecklist(category: "Departure")
-        clearChecklist(category: "Arrival")
-    }
-    
-    func clearChecklist(category: String?) {
-        for index in 0..<modelData.checklist.count {
-            if modelData.checklist[index].category == category {
-                print("Clearing item")
-                modelData.checklist[index].isDone = false
             }
         }
     }
